// src/app/models/IndustrializationPlan.ts
export interface IndustrializationPlan {
    id?: number;
    project: string;
    customer_1: string;
    derivative_1: string;
    maxKapazitaetLinieFzgTag: number;
    productionConceptResponsible: string;
    drawingStatus: string;
    Date: Date;
    aSampleDate: Date;
    bSampleDate: Date;
    cSampleDate: Date;
    sopDate: Date;
    plant: string;
    sverweis: string;
    equipment: string;
    Derivative: string;
    usedInIndustrialization: string;
    newInvest: string;
    relocation: string;
    availableInPlant: string;
    pricePerPiece: number;
    qty: number;
    sum: number;
    pamNumber: string;
    pamRequestDate: Date;
    pamReleaseDate: Date;
    daysForPam: number;
    sopMinus2Invest: number;
    sopMinus1Invest: number;
    sopInvest: number;
    sopPlus1Invest: number;
    VerantWortlich: string;
    orderNumber: number;
    budgetReleased: number;
    orderPrice: number;
    orderDate: Date;
    latestOrderDate: Date;
    leadTimeWeeks: number;
    transportationTimeWeeks: number;
    installationTimeWeeks: number;
    targetDate: Date;
    realDeliveryDate: Date;
    confirmedDeliveryDate: Date;
    projectPhase: string;
    Comment: string;
  }